+++ packages/micropip/src/micropip/_compat_in_pyodide.py
@@ -1,20 +1,12 @@
-from pathlib import Path
-
 from pyodide._core import IN_BROWSER
 from pyodide.http import pyfetch
 
 try:
     import pyodide_js
-    from pyodide_js import loadedPackages
+    from pyodide_js import loadedPackages, loadPackage
+    from pyodide_js._api import loadDynlib  # type: ignore[import]
 
     BUILTIN_PACKAGES = pyodide_js._api.packages.to_py()
-
-    # Random note: getsitepackages is not available in a virtual environment...
-    # See https://github.com/pypa/virtualenv/issues/228 (issue is closed but
-    # problem is not fixed)
-    from site import getsitepackages
-
-    WHEEL_BASE = Path(getsitepackages()[0])
 except ImportError:
     if IN_BROWSER:
         raise
@@ -32,8 +24,8 @@ async def fetch_string(url: str, kwargs: dict[str, str]) -> str:
 __all__ = [
     "fetch_bytes",
     "fetch_string",
-    "WHEEL_BASE",
     "BUILTIN_PACKAGES",
     "loadedPackages",
-    "pyodide_js",
+    "loadDynlib",
+    "loadPackage",
 ]