+++ packages/micropip/src/micropip/_compat_not_in_pyodide.py
@@ -1,9 +1,5 @@
-import tempfile
-from pathlib import Path
 from typing import Any
 
-# Provide stubs for testing in native python
-WHEEL_BASE = Path(tempfile.mkdtemp())
 BUILTIN_PACKAGES: dict[str, dict[str, Any]] = {}
 
 
@@ -24,19 +20,24 @@ async def fetch_string(url: str, kwargs: dict[str, str]) -> str:
     return (await fetch_bytes(url, kwargs)).decode()
 
 
+async def loadDynlib(dynlib: str, is_shared_lib: bool) -> None:
+    pass
+
+
 class pyodide_js_:
     def __get__(self, attr):
         raise RuntimeError(f"Attempted to access property '{attr}' on pyodide_js dummy")
 
 
-pyodide_js: Any = pyodide_js_()
+def loadPackage(packages: str | list[str]) -> None:
+    pass
 
 
 __all__ = [
+    "loadDynlib",
     "fetch_bytes",
     "fetch_string",
-    "WHEEL_BASE",
     "BUILTIN_PACKAGES",
     "loadedPackages",
-    "pyodide_js",
+    "loadPackage",
 ]