+++ packages/pywavelets/test_pywt.py
@@ -2,8 +2,9 @@ import pytest
 from pyodide_test_runner import run_in_pyodide
 
 
+@pytest.mark.driver_timeout(30)
 @pytest.mark.xfail_browsers(chrome="xfail")
-@run_in_pyodide(packages=["pywavelets"], driver_timeout=30)
+@run_in_pyodide(packages=["pywavelets"])
 def test_pywt(selenium):
     import numpy as np
     import pywt