+++ packages/scipy/test_scipy.py
@@ -1,12 +1,9 @@
+import pytest
 from pyodide_test_runner import run_in_pyodide
 
-run_in_pyodide_scipy = run_in_pyodide(
-    packages=["scipy"],
-    driver_timeout=40,
-)
 
-
-@run_in_pyodide_scipy
+@pytest.mark.driver_timeout(40)
+@run_in_pyodide(packages=["scipy"])
 def test_scipy_linalg(selenium):
     import numpy as np
     import scipy.linalg
@@ -22,27 +19,31 @@ def test_scipy_linalg(selenium):
     assert_allclose(res, np.identity(N), rtol=1e-07, atol=1e-9)
 
 
-@run_in_pyodide_scipy
+@pytest.mark.driver_timeout(40)
+@run_in_pyodide(packages=["scipy"])
 def test_brentq(selenium):
     from scipy.optimize import brentq
 
     brentq(lambda x: x, -1, 1)
 
 
-@run_in_pyodide_scipy
+@pytest.mark.driver_timeout(40)
+@run_in_pyodide(packages=["scipy"])
 def test_dlamch(selenium):
     from scipy.linalg import lapack
 
     lapack.dlamch("Epsilon-Machine")
 
 
-@run_in_pyodide_scipy
+@pytest.mark.driver_timeout(40)
+@run_in_pyodide(packages=["scipy"])
 def test_binom_ppf(selenium):
     from scipy.stats import binom
 
     assert binom.ppf(0.9, 1000, 0.1) == 112
 
 
+@pytest.mark.driver_timeout(40)
 @run_in_pyodide(packages=["pytest", "scipy-tests"])
 def test_scipy_pytest(selenium):
     import pytest