+++ pyodide-test-runner/pyodide_test_runner/fixture.py
@@ -2,13 +2,59 @@ import contextlib
 
 import pytest
 
-from .browser import ChromeWrapper, FirefoxWrapper, NodeWrapper, SeleniumWrapper
+from .browser import (
+    BrowserWrapper,
+    NodeWrapper,
+    PlaywrightChromeWrapper,
+    PlaywrightFirefoxWrapper,
+    SeleniumChromeWrapper,
+    SeleniumFirefoxWrapper,
+)
 from .server import spawn_web_server
 from .utils import parse_driver_timeout, set_webdriver_script_timeout
 
 
+@pytest.fixture(scope="module")
+def playwright_browsers(request):
+    if request.config.option.runner.lower() != "playwright":
+        yield {}
+    else:
+        # import playwright here to allow running tests without playwright installation
+        try:
+            from playwright.sync_api import sync_playwright
+        except ImportError:
+            pytest.exit(
+                "playwright not installed. try `pip install playwright && python -m playwright install`",
+                returncode=1,
+            )
+
+        with sync_playwright() as p:
+            try:
+                chromium = p.chromium.launch(
+                    args=[
+                        "--js-flags=--expose-gc",
+                    ],
+                )
+                firefox = p.firefox.launch()
+                # webkit = p.webkit.launch()
+            except Exception as e:
+                pytest.exit(f"playwright failed to launch\n{e}", returncode=1)
+            try:
+                yield {
+                    "chrome": chromium,
+                    "firefox": firefox,
+                    # "webkit": webkit,
+                }
+            finally:
+                chromium.close()
+                firefox.close()
+                # webkit.close()
+
+
 @contextlib.contextmanager
-def selenium_common(request, web_server_main, load_pyodide=True, script_type="classic"):
+def selenium_common(
+    request, web_server_main, load_pyodide=True, script_type="classic", browsers=None
+):
     """Returns an initialized selenium object.
 
     If `_should_skip_test` indicate that the test will be skipped,
@@ -16,34 +62,45 @@ def selenium_common(request, web_server_main, load_pyodide=True, script_type="cl
     """
 
     server_hostname, server_port, server_log = web_server_main
-    cls: type[SeleniumWrapper]
-    if request.param == "firefox":
-        cls = FirefoxWrapper
-    elif request.param == "chrome":
-        cls = ChromeWrapper
-    elif request.param == "node":
-        cls = NodeWrapper
-    else:
-        raise AssertionError(f"Unknown browser: {request.param}")
+    runner_type = request.config.option.runner.lower()
+    cls: type[BrowserWrapper]
+
+    browser_set = {
+        ("selenium", "firefox"): SeleniumFirefoxWrapper,
+        ("selenium", "chrome"): SeleniumChromeWrapper,
+        ("selenium", "node"): NodeWrapper,
+        ("playwright", "firefox"): PlaywrightFirefoxWrapper,
+        ("playwright", "chrome"): PlaywrightChromeWrapper,
+        ("playwright", "node"): NodeWrapper,
+    }
+
+    cls = browser_set.get((runner_type, request.param))
+    if cls is None:
+        raise AssertionError(
+            f"Unknown runner or browser: {runner_type} / {request.param}"
+        )
 
     dist_dir = request.config.getoption("--dist-dir")
-    selenium = cls(
+    runner = cls(
         server_port=server_port,
         server_hostname=server_hostname,
         server_log=server_log,
         load_pyodide=load_pyodide,
+        browsers=browsers,
         script_type=script_type,
         dist_dir=dist_dir,
     )
     try:
-        yield selenium
+        yield runner
     finally:
-        selenium.quit()
+        runner.quit()
 
 
 @pytest.fixture(params=["firefox", "chrome", "node"], scope="function")
-def selenium_standalone(request, web_server_main):
-    with selenium_common(request, web_server_main) as selenium:
+def selenium_standalone(request, web_server_main, playwright_browsers):
+    with selenium_common(
+        request, web_server_main, browsers=playwright_browsers
+    ) as selenium:
         with set_webdriver_script_timeout(
             selenium, script_timeout=parse_driver_timeout(request.node)
         ):
@@ -54,9 +111,13 @@ def selenium_standalone(request, web_server_main):
 
 
 @pytest.fixture(params=["firefox", "chrome", "node"], scope="module")
-def selenium_esm(request, web_server_main):
+def selenium_esm(request, web_server_main, playwright_browsers):
     with selenium_common(
-        request, web_server_main, load_pyodide=True, script_type="module"
+        request,
+        web_server_main,
+        load_pyodide=True,
+        browsers=playwright_browsers,
+        script_type="module",
     ) as selenium:
         with set_webdriver_script_timeout(
             selenium, script_timeout=parse_driver_timeout(request.node)
@@ -68,9 +129,15 @@ def selenium_esm(request, web_server_main):
 
 
 @contextlib.contextmanager
-def selenium_standalone_noload_common(request, web_server_main, script_type="classic"):
+def selenium_standalone_noload_common(
+    request, web_server_main, playwright_browsers, script_type="classic"
+):
     with selenium_common(
-        request, web_server_main, load_pyodide=False, script_type=script_type
+        request,
+        web_server_main,
+        load_pyodide=False,
+        browsers=playwright_browsers,
+        script_type=script_type,
     ) as selenium:
         with set_webdriver_script_timeout(
             selenium, script_timeout=parse_driver_timeout(request.node)
@@ -82,35 +149,36 @@ def selenium_standalone_noload_common(request, web_server_main, script_type="cla
 
 
 @pytest.fixture(params=["firefox", "chrome"], scope="function")
-def selenium_webworker_standalone(request, web_server_main, script_type):
+def selenium_webworker_standalone(
+    request, web_server_main, playwright_browsers, script_type
+):
     # Avoid loading the fixture if the test is going to be skipped
     if request.param == "firefox" and script_type == "module":
         pytest.skip("firefox does not support module type web worker")
 
     with selenium_standalone_noload_common(
-        request, web_server_main, script_type=script_type
+        request, web_server_main, playwright_browsers, script_type=script_type
     ) as selenium:
         yield selenium
 
 
-@pytest.fixture(params=["classic", "module"], scope="module")
-def script_type(request):
-    return request.param
-
-
 @pytest.fixture(params=["firefox", "chrome", "node"], scope="function")
-def selenium_standalone_noload(request, web_server_main):
+def selenium_standalone_noload(request, web_server_main, playwright_browsers):
     """Only difference between this and selenium_webworker_standalone is that
     this also tests on node."""
 
-    with selenium_standalone_noload_common(request, web_server_main) as selenium:
+    with selenium_standalone_noload_common(
+        request, web_server_main, playwright_browsers
+    ) as selenium:
         yield selenium
 
 
 # selenium instance cached at the module level
 @pytest.fixture(params=["firefox", "chrome", "node"], scope="module")
-def selenium_module_scope(request, web_server_main):
-    with selenium_common(request, web_server_main) as selenium:
+def selenium_module_scope(request, web_server_main, playwright_browsers):
+    with selenium_common(
+        request, web_server_main, browsers=playwright_browsers
+    ) as selenium:
         yield selenium
 
 
@@ -135,6 +203,21 @@ def selenium(request, selenium_module_scope):
             yield selenium
 
 
+@pytest.fixture(params=["firefox", "chrome"], scope="function")
+def console_html_fixture(request, web_server_main, playwright_browsers):
+    with selenium_common(
+        request, web_server_main, load_pyodide=False, browsers=playwright_browsers
+    ) as selenium:
+        selenium.goto(
+            f"http://{selenium.server_hostname}:{selenium.server_port}/console.html"
+        )
+        selenium.javascript_setup()
+        try:
+            yield selenium
+        finally:
+            print(selenium.logs)
+
+
 @pytest.fixture(scope="session")
 def web_server_main(request):
     """Web server that serves files in the dist directory"""
@@ -147,3 +230,8 @@ def web_server_secondary(request):
     """Secondary web server that serves files dist directory"""
     with spawn_web_server(request.config.option.dist_dir) as output:
         yield output
+
+
+@pytest.fixture(params=["classic", "module"], scope="module")
+def script_type(request):
+    return request.param