+++ src/tests/test_pyproxy.py
@@ -369,10 +369,10 @@ def test_pyproxy_get_buffer_type_argument(selenium, array_type):
         if fmt.lower() == "q":
             assert result == [hex(x).replace("0x", "") for x in list(mv.cast(fmt))]
         elif fmt == "f" or fmt == "d":
-            from math import isclose
+            from math import isclose, isnan
 
             for a, b in zip(result, list(mv.cast(fmt))):
-                if a and b:
+                if a and b and not (isnan(a) or isnan(b)):
                     assert isclose(a, b)
         else:
             assert result == list(mv.cast(fmt))